Expanded version number ↔️
Our CI process added a build number to the end of the text in the Version node. Though it was a build number, it ended up being represented as a bugfix version number. This shifts the build number to the 4th place and therefore the proper slot for a build number.